Add files via upload
This directory contains the initial version of the JavaScript Image Manager Dynamic Model Website. The files in this directory provide the basic structure of the website, but are not fully furnished yet. No galleries or images are displayed yet. This version is intended to serve as a starting point for the development of the website.